requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.7
coverage==7.10.2
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
packaging==25.0
pandas==2.1.4
pluggy==1.6.0
Pygments==2.19.2
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
rich==13.7.0
six==1.17.0
tomli==2.2.1
typer==0.9.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
xmltodict==0.13.0
